SFX-248: Remove module property from package.json (#3)
This conflicts with Webpack when the package is inside node_modules.

Also ensure that build files are included when publishing.

The "dist" and "esnext" directories are excluded by .gitignore and would
therefore be excluded when publishing. Adding them to the "files" field
in package.json ensures that they are included.